Migrated to modern PaperCommandManager
