+++ platforms/bukkit/common/src/main/java/com/dfsek/terra/bukkit/TerraBukkitPlugin.java
@@ -23,10 +23,8 @@ import org.bukkit.Bukkit;
 import org.bukkit.generator.ChunkGenerator;
 import org.bukkit.plugin.java.JavaPlugin;
 import org.incendo.cloud.SenderMapper;
-import org.incendo.cloud.brigadier.CloudBrigadierManager;
-import org.incendo.cloud.bukkit.CloudBukkitCapabilities;
 import org.incendo.cloud.execution.ExecutionCoordinator;
-import org.incendo.cloud.paper.LegacyPaperCommandManager;
+import org.incendo.cloud.paper.PaperCommandManager;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.slf4j.Logger;
@@ -73,7 +71,7 @@ public class TerraBukkitPlugin extends JavaPlugin {
         platform.getEventManager().callEvent(new PlatformInitializationEvent());
 
         try {
-            LegacyPaperCommandManager<CommandSender> commandManager = getCommandSenderPaperCommandManager();
+            PaperCommandManager<CommandSender> commandManager = getCommandSenderPaperCommandManager();
 
             platform.getEventManager().callEvent(new CommandRegistrationEvent(commandManager));
 
@@ -93,25 +91,15 @@ public class TerraBukkitPlugin extends JavaPlugin {
     }
 
     @NotNull
-    private LegacyPaperCommandManager<CommandSender> getCommandSenderPaperCommandManager() throws Exception {
-        // TODO: Update to PaperCommandManager
-        LegacyPaperCommandManager<CommandSender> commandManager = new LegacyPaperCommandManager<>(
-            this,
-            ExecutionCoordinator.simpleCoordinator(),
-            SenderMapper.create(
+    private PaperCommandManager<CommandSender> getCommandSenderPaperCommandManager() throws Exception {
+        PaperCommandManager<CommandSender> commandManager = PaperCommandManager.builder(SenderMapper.create(
                 BukkitAdapter::adapt,
                 BukkitAdapter::adapt
-            ));
+            ))
+            .executionCoordinator(ExecutionCoordinator.simpleCoordinator())
+            .buildOnEnable(this);
 
-        if(commandManager.hasCapability(CloudBukkitCapabilities.NATIVE_BRIGADIER)) {
-            commandManager.registerBrigadier();
-            final CloudBrigadierManager<?, ?> brigManager = commandManager.brigadierManager();
-            if(brigManager != null) {
-                brigManager.setNativeNumberSuggestions(false);
-            }
-        } else if(commandManager.hasCapability(CloudBukkitCapabilities.ASYNCHRONOUS_COMPLETION)) {
-            commandManager.registerAsynchronousCompletions();
-        }
+        commandManager.brigadierManager().setNativeNumberSuggestions(false);
 
         return commandManager;
     }
